1.21.11 port; to adapt new version; bug fix

+++ common/src/main/java/com/kwwsyk/endinv/common/client/gui/bg/ScreenRectangleWidgetParam.java
@@ -3,7 +3,7 @@ package com.kwwsyk.endinv.common.client.gui.bg;
 public record ScreenRectangleWidgetParam(int x, int y, int width, int height) implements IRectangleParam {
 
     @Override
-    public boolean hasClickedOn(int mouseX, int mouseY){
+    public boolean hasClickedOn(double mouseX, double mouseY){
         return mouseX>= x && mouseX<= x + width && mouseY>= y && mouseY<= y + height;
     }
 }
